Extracts kitty commands to an image type that can be used for explicit positioning

+++ src/window_ctrl.rs
@@ -17,22 +17,22 @@ impl Error for WindowCtrlError {}
 
 #[derive(Debug)]
 pub struct WindowSize {
-    pub rows: u16,
-    pub cols: u16,
-    pub x_pix: u16,
-    pub y_pix: u16,
-    pub pix_per_row: f32,
-    pub pix_per_col: f32,
+    pub rows: u32,
+    pub cols: u32,
+    pub x_pix: u32,
+    pub y_pix: u32,
+    pub pix_per_row: u32,
+    pub pix_per_col: u32,
 }
 
 impl WindowSize {
     fn build_from_ioctl(ioctl_data: [u16; 4]) -> WindowSize {
-        let rows = ioctl_data[0];
-        let cols = ioctl_data[1];
-        let x_pix = ioctl_data[2];
-        let y_pix = ioctl_data[3];
-        let pix_per_col = x_pix as f32 / cols as f32;
-        let pix_per_row = y_pix as f32 / rows as f32;
+        let rows = ioctl_data[0] as u32;
+        let cols = ioctl_data[1] as u32;
+        let x_pix = ioctl_data[2] as u32;
+        let y_pix = ioctl_data[3] as u32;
+        let pix_per_col = x_pix / cols;
+        let pix_per_row = y_pix / rows;
         WindowSize {
             rows,
             cols,
